# Schedule 'force_special_times=True' performance optimization