--- a/tests/test_exchange_calendar_cme_globex_crypto.py
+++ b/tests/test_exchange_calendar_cme_globex_crypto.py
@@ -21,7 +21,7 @@ def test_is_different():
     late = cal.late_opens(sched).index
     assert late.empty
 
-    sched.loc["2020-01-20", "market_open"] += pd.Timedelta("3H")
+    sched.loc["2020-01-20", "market_open"] += pd.Timedelta("3h")
     late = cal.late_opens(sched).index
     assert late.shape == (1,) and late[0] == pd.Timestamp("2020-01-20")