--- a/tests/test_market_calendar.py
+++ b/tests/test_market_calendar.py
@@ -324,7 +324,7 @@ def test_change_add_remove_time():
         cal.add_time("wrong_format", (-1, time(10)))
 
     with pytest.raises(AssertionError):
-        cal.add_time("wrong_format", pd.Timedelta("5H"))
+        cal.add_time("wrong_format", pd.Timedelta("5h"))
 
 
 def test_add_change_remove_time_w_open_close_map():
@@ -1464,7 +1464,7 @@ def test_open_at_time_interruptions():
     assert cal.open_at_time(sched, "2010-01-12 16:57:00", include_close=True) is False
 
     # interruption between market_close/post
-    sched.iloc[2, [-2, -1]] += pd.Timedelta("1H")
+    sched.iloc[2, [-2, -1]] += pd.Timedelta("1h")
     assert cal.open_at_time(sched, "2010-01-12 17:56:00") is False
     assert cal.open_at_time(sched, "2010-01-12 17:55:00", include_close=True) is False
     assert cal.open_at_time(sched, "2010-01-12 17:57:00", include_close=True) is False