When force_special_times is True, the special_times are applied to other columns by editing an intermediate pandas Series Object. This was changed so that the code now edits an intermediate numpy array which is slightly more performant.

This change is most noticeable when generating longer schedules, or calendars that have re-occurring special times (such as XTAE's early sunday close)


### Average NYSE Schedule Generation Time [2000-01-01 -> 2025-01-01]
|  | force_special_times = True | False | None |
| -------- | ------- | ------- | ------- |
| No Optimizations | 214.3ms (+38%) | 173.0ms (+14%) | 121.6ms (~0%*)
| _tryholidays() Optimization (PR #357) | 188.0ms (+21%) | 152.8ms (~0%*) | 118.6ms (~0%*)
| _tryHolidays() & _adj_others Optimization | 155.4ms ( - ) | 151.5ms ( - ) | 122.8ms ( - )

\* These numbers should, in theory, be identical to their values in the last row. Random variations in execution time are causing the small discrepancies