@jack-of-some-trades 2024-12-14
Being a Hardware Engineer I'm pretty new to Github. What I pull-requested
was an isolated update, but I'm still making other updates to the repo to
add a couple things. Because of that, I just committed some partial changes
to my own Dev Branch. I didn't realise that those automatically got roped
in with the already open PR. Once I'm done adding stuff, and everything is
in a finalized state then I was going to reopen the PR.

On Sat, Dec 14, 2024 at 11:59 AM rsheftel ***@***.***> wrote:

> I see you closed this without merging, is there a reason for that?
>
> —
> Reply to this email directly, view it on GitHub
> <https://github.com/rsheftel/pandas_market_calendars/pull/358#issuecomment-2543211569>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/BFIBUJ2K6SRLUOQAYRNVP6T2FRWXRAVCNFSM6AAAAABTOL7VS6VHI2DSMVQWIX3LMV43OSLTON2WKQ3PNVWWK3TUHMZDKNBTGIYTCNJWHE>
> .
> You are receiving this because you modified the open/close state.Message
> ID: ***@***.***>
>