@jack-of-some-trades 2025-01-06
No problem! There are still a few other (hopefully smaller) things along the same vein i might add here soon. They're things I personally could use, but fit well in the scope of your library so I figured I'd add them as contributions instead.