@rsheftel 2025-01-06
Please feel free to add anything you think may be useful.

On Sun, Jan 5, 2025, at 7:03 PM, Bryce Hawk wrote:
> 
> 
> No problem! There are still a few other (hopefully smaller) things along the same vein i might add here soon. They're things I personally could use, but fit well in the scope of your library so I figured I'd add them as contributions instead.
> 
> 
> —
> Reply to this email directly, view it on GitHub <https://github.com/rsheftel/pandas_market_calendars/pull/358#issuecomment-2571800624>, or unsubscribe <https://github.com/notifications/unsubscribe-auth/ACYBJI6TY4YOT3D5QEPZY632JHB4NAVCNFSM6AAAAABTOL7VS6VHI2DSMVQWIX3LMV43OSLTON2WKQ3PNVWWK3TUHMZDKNZRHAYDANRSGQ>.
> You are receiving this because you modified the open/close state.Message ID: ***@***.***>
>